requirements.txt
aiohttp>=3.5.4
async-timeout>=3.0.1
asyncio>=3.4.3
attrs>=18.2.0
cffi>=1.12.1
chardet==3.0.4
discord.py>=1.2.3[voice]
idna==2.8
multidict==4.5.2
psutil>=5.5.1
Layer-7-Utilities>=1.2.11rc10
pycparser==2.19
PyNaCl>=1.2.1
pytest
six==1.12.0
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
SQLAlchemy>=1.2.18
websockets==6.0
yarl==1.3.0
parsedatetime==2.4
py-dateutil>=2.2
asyncpg>=0.18.3
sqlalchemy-citext>=1.3.0
sentry_sdk>=0.12.3
redis>=3.3.8
num2words>=0.5.10
